fix: noweeb style checks
