+++ package.json
@@ -87,7 +87,7 @@
     "react-native-carplay": "2.4.1-beta.0",
     "react-native-device-info": "^13.0.0",
     "react-native-draggable-flatlist": "^4.0.1",
-    "react-native-gesture-handler": "2.19.0",
+    "react-native-gesture-handler": "2.20.0",
     "react-native-get-random-values": "^1.11.0",
     "react-native-image-colors": "^2.4.0",
     "react-native-lyric": "git+https://github.com/lovegaoshi/react-native-lyric.git#dev",
@@ -135,8 +135,8 @@
     "@types/he": "^1.2.3",
     "@types/jest": "^29.5.13",
     "@types/md5": "^2.3.5",
-    "@types/node": "^22.6.1",
-    "@types/react": "~18.3.9",
+    "@types/node": "^22.7.4",
+    "@types/react": "~18.3.10",
     "@types/react-native": "^0.73.0",
     "@types/react-native-background-timer": "^2.0.2",
     "@types/react-native-share-menu": "^5.0.5",
@@ -158,7 +158,7 @@
     "eslint-plugin-import": "^2.30.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^5.2.1",
-    "eslint-plugin-react": "^7.36.1",
+    "eslint-plugin-react": "^7.37.0",
     "eslint-plugin-react-hooks": "^4.6.2",
     "gts": "^5.3.1",
     "jest": "^29.7.0",
