+++ src/components/dialogs/NoxInput.tsx
@@ -1,15 +1,20 @@
 import React from 'react';
+
 import { TextInput } from 'react-native-paper';
+import { TextStyle, TextInput as RNTextInput } from 'react-native';
+
 import { useNoxSetting } from '@stores/useApp';
 
 interface InputProps {
   handleSubmit?: () => void;
-  label: string;
+  label?: string;
   autofocus?: boolean;
   selectTextOnFocus?: boolean;
   text: string;
   setText: (text: string) => void;
   secureTextEntry?: boolean;
+  style?: TextStyle;
+  reactNative?: boolean;
 }
 
 export default ({
@@ -20,14 +25,18 @@ export default ({
   text,
   setText,
   secureTextEntry,
+  style,
+  reactNative = false,
 }: InputProps) => {
   const playerStyle = useNoxSetting(state => state.playerStyle);
+  const Input = reactNative ? RNTextInput : TextInput;
 
   return (
-    <TextInput
-      label={label}
+    <Input
+      style={style}
       value={text}
       onChangeText={(val: string) => setText(val)}
+      label={label}
       onSubmitEditing={handleSubmit}
       selectTextOnFocus={selectTextOnFocus}
       selectionColor={playerStyle.customColors.textInputSelectionColor}
